+++ src/test/scala/org/podval/tools/testing/Platform.scala
@@ -1,6 +1,6 @@
 package org.podval.tools.testing
 
-import org.podval.tools.build.{Dependency, JavaDependency, ScalaDependency, ScalaLibrary, Version}
+import org.podval.tools.build.{Dependency, JavaDependency, ScalaDependency, ScalaLibraryDependency, Version}
 import org.podval.tools.node.NodeDependency
 import org.podval.tools.testing.framework.FrameworkDescriptor
 
@@ -10,9 +10,7 @@ final class Platform(
   val nodeVersion: Version = NodeDependency.versionDefault
 ):
   def displayName: String = s"""in Scala v$scalaVersion${if isScalaJS then " with ScalaJS" else ""}"""
-
-  def scalaLibraryDependency: Dependency.WithVersion = ScalaLibrary.Scala.forVersion(scalaVersion)
-
+  
   def getNodeVersion: Option[Version] = if !isScalaJS then None else Some(nodeVersion)
   
   def toDependency(framework: FrameworkDescriptor): Dependency.WithVersion =
@@ -21,8 +19,14 @@ final class Platform(
     val group: String = framework.group
     val artifact: String = framework.artifact
 
-    (if !framework.isScalaDependency then JavaDependency(group, artifact) else (
-      if scalaVersion.major == ScalaLibrary.Scala3.versionMajor
-      then ScalaDependency.Scala3(group, artifact, isScalaJS = isScalaJS)
-      else ScalaDependency.Scala2(group, artifact, isScalaJS = isScalaJS)
-    ).withScalaVersion(scalaVersion)).withVersion(Version(framework.versionDefault))
+    val result: Dependency = 
+      if !framework.isScalaDependency
+      then JavaDependency(group, artifact)
+      else ScalaDependency.mk(
+        scalaVersion,
+        group, 
+        artifact,
+        isScalaJS
+      ).withScalaVersion(scalaVersion)
+      
+    result.withVersion(Version(framework.versionDefault))