+++ src/test/scala/org/podval/tools/testing/ScalaJSTutorialTest.scala
@@ -1,15 +1,15 @@
 package org.podval.tools.testing
 
-import org.podval.tools.build.ScalaLibrary
+import org.podval.tools.build.ScalaLibraryDependency
 
 class ScalaJSTutorialTest extends GroupingFunSpec:
   groupTest(
     features = Seq(Feature("ScalaJS Tutorial")),
     fixtures = Seq(ScalaJSTutorialScalaTestFixture),
     platforms = Seq(
-      Platform(ScalaLibrary.Scala3.versionDefault  , isScalaJS = true),
-      Platform(ScalaLibrary.Scala2.versionDefault13, isScalaJS = true),
-      Platform(ScalaLibrary.Scala2.versionDefault12, isScalaJS = true),
+      Platform(ScalaLibraryDependency.Scala3.versionDefault  , isScalaJS = true),
+      Platform(ScalaLibraryDependency.Scala2.versionDefault13, isScalaJS = true),
+      Platform(ScalaLibraryDependency.Scala2.versionDefault12, isScalaJS = true),
     ),
     groupByFeature = true,
     combinedFixtureNameOpt = None